feat: first pass support for bind mounts
